nexus/requirements.txt
boto3
python-dotenv
requests
moviepy
google-api-python-client
google-auth
google-auth-oauthlib
elevenlabs
supabase
ffmpeg-python
discord-webhook
Pillow
